store theme in redux
