+++ components/Header/Operations.module.css
@@ -1,0 +1,6 @@
+.operations__container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    height: 100%;
+}
