+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import type { AppProps } from "next/app";
 import { CacheProvider, EmotionCache } from "@emotion/react";
 import { ThemeProvider, CssBaseline, createTheme } from "@mui/material";
-import { Provider as ReduxProvider } from "react-redux";
+import { Provider as ReduxProvider, useSelector } from "react-redux";
 import { Hydrate, QueryClientProvider } from "react-query";
 import { queryClient } from "../src/api";
 
@@ -13,6 +13,7 @@ import darkThemeOptions from "../styles/theme/darkThemeOptions";
 import "../styles/globals.css";
 import { useEffect, useState } from "react";
 import { wrapper } from "../src/store/store";
+import { selectTheme } from "../src/store/themeSlice";
 
 interface MyAppProps extends AppProps {
   emotionCache?: EmotionCache;
@@ -24,44 +25,36 @@ const clientSideEmotionCache = createEmotionCache();
 const lightTheme = createTheme(lightThemeOptions);
 const darkTheme = createTheme(darkThemeOptions);
 
-function getActiveTheme(themeMode: "light" | "dark") {
-  return themeMode === "light" ? lightTheme : darkTheme;
-}
-
 const MyApp: React.FunctionComponent<MyAppProps> = ({ Component, ...rest }) => {
   const { store, props } = wrapper.useWrappedStore(rest);
   const { emotionCache = clientSideEmotionCache, pageProps } = props;
-
-  const [activeTheme, setActiveTheme] = useState(lightTheme);
-  const [selectedTheme, setSelectedTheme] = useState<"light" | "dark">("light");
-
-  const toggleTheme: React.MouseEventHandler<HTMLAnchorElement> = () => {
-    const desiredTheme = selectedTheme === "light" ? "dark" : "light";
-    setSelectedTheme(desiredTheme);
-  };
-
-  useEffect(() => {
-    setActiveTheme(getActiveTheme(selectedTheme));
-    console.log("theme switched: ", selectedTheme);
-  }, [selectedTheme]);
 
   return (
     <ReduxProvider store={store}>
       <QueryClientProvider client={queryClient}>
         <Hydrate state={pageProps.dehydratedState}>
           <CacheProvider value={emotionCache}>
-            <ThemeProvider theme={activeTheme}>
-              <CssBaseline />
-              <Component
-                {...pageProps}
-                toggleTheme={toggleTheme}
-                currentTheme={selectedTheme}
-              />
-            </ThemeProvider>
+            <ThemeWrapper>
+              <Component {...pageProps} />
+            </ThemeWrapper>
           </CacheProvider>
         </Hydrate>
       </QueryClientProvider>
     </ReduxProvider>
+  );
+};
+
+// 这里有一些注意事项:
+// 我们将 theme 这个参数存储进了 redux, 然后在ThemeProvider中取出这个参数。
+// 但是，如果Redux Provider和ThemeProvider在同一个元素中，那么ThemeProvider无法获取Redux Context中的数据。
+// 这时候就需要一个中间组件，将Redux Provider和ThemeProvider分开。
+const ThemeWrapper = (props) => {
+  const themeState = useSelector(selectTheme);
+  return (
+    <ThemeProvider theme={themeState ? lightTheme : darkTheme}>
+      <CssBaseline />
+      {props.children}
+    </ThemeProvider>
   );
 };
