+++ pages/home.module.css
@@ -105,10 +105,7 @@
 }
 
 .header__rightside {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    height: 100%;
+
 }
 
 .author__info,
